* Moved screenshot functionality from FrameBuffersManager to PostProcessor. PostProcessor however still gets the GPU data from the FrameBuffersManager.
